requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.6.1
dj-database-url==1.1.0
django==4.2.24
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.7
gunicorn==22.0.0
importlib-resources==6.4.5
inflection==0.5.1
jq==1.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.5.4
packaging==25.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
dj-database-url<1.2.0